pc/control: Only accept/use session in Create
The constructors were all taking a session but only using them in
Create. Might as well only take in Create.

History: The session argument used to be used in Mutate as well, but
that was removed in #524. After this point, Create is the only user of a
session, so it can be the only function that takes one.