Add pod cluster identifying info to RUCompletionEventDetails
Previously audit log records that were created to mark the cancellation
or successful completion of a rolling update did not contain information
to identify the pod cluster the RU belonged to. This commit adds that
information because audit log systems will likely want to be able to
look up this information based on pod cluster.