Fix RemoveAllLabelsTxn in consul applicator.
The intention was that that call wouldn't actually remove labels, it
would just add an api.KVTxnOp to the passed transaction within the
passed context to do so, and the actual deletion wouldn't occur until
the transaction was committed. However the code had the exact same
behavior as the synchronous deletion in RemoveAllLabels().

This commit reimplements the function as intended and adds a regression
test.