Do not consider a missing implicit entry point an error.
This change is intended to cut down on error noise in p2-preparer. The
hoist artifact spec historically required all "entry points" for an
artifact to be located at bin/launch whether it's a directory of entry
points or a file. At a later point in time the pod manifest spec was
extended to allow specifying alternate entry points, at which point it
was considered an error if an entry point is missing.

However it is not uncommon for an artifact to have no entry points, for
instance when the intention is just to deploy some scripts to a server
and not run any processes. These cases now generate errors which pollute
preparer logs and exception systems.

This commit tries to take the best of both worlds by considering a
missing entry point to be an error only if it was explicitly enumerated
in the pod manifest. If the implicit default one is missing, it is no
longer an error.