roll farm: only delete old RC if replica count is zero. otherwise page.
Previously the RU farm would always zero out the old RC's replica count
and then delete it if the "LeaveOld" flag was not set. This can lead to
bad situations if a math error is encountered when creating an RU
because we might delete an RC that was managing pods and then those pods
become orphaned.

This commit causes the RU farm to only delete the old RC if its replica
count is zero. Otherwise it will send an alert so a human user can
correct the situation.