travis: Run once with the race detector and once without
We're working around https://github.com/square/p2/issues/832

We're refining the method used in https://github.com/square/p2/pull/831
It becomes harder to maintain the list of packages that will get
excluded from the race detector (and thus must be rerun without it), so
we will just run all tests without and all tests with.

Since we desire not to slow down the tests by running them twice, we
hope to mitigate the slowdown by running the two in parallel.