Add details schema for "rc retargeting events".
An rc retargeting event represents an RC changing the set of nodes it
targets, e.g. due to a replica count change, selector change, or a
change to the labels of a node.

This commit creates the json schema for the details for an event as well
as a constructor for easily creating the json.RawMessage for use in an
audit log record.