Make replication controllers use transactions to schedule nodes.
This means that pods will be scheduled along with their labels in a
transaction which removes failure modes where labels will be out of sync
with scheduled pods due to consul server availability interruptions.

This also paves the way to creating audit log records as part of an RC
retargeting event.