require availability zone and cluster name pod labels on RCs
RCs have a "pod_labels" field which represents a set of labels that
should be applied to every pod that is scheduled by that RC. Previously
there has been no requirement that that set of labels includes the
availability_zone and cluster_name to identify the pod cluster that the
RC and its pods are members of. pod_id can be inferred from the pod
manifest and is already automatically set.

This commit enforces that the set of labels have availability zone and
cluster name in anticipation of having RCs create audit log records
every time the set of nodes they target changes. This is because audit
log records are linked to pod clusters, and as a result RCs must be as
well.