Fix the way labels are created for RCs in rcstore.CreateTxn().
Typically transactional store operations (that take a context.Context
that has a transaction value embedded in it) will not actually perform
any network requests to consul but will instead append consul operations
to the context to be applied when the caller calls Commit() on the
transaction. This is useful so callers can string together arbitrary kv
operations and apply them all at once in a transaction.

The rcstore was violating this practice in its CreateTxn() function. It
would do the right thing when creating the RC by just adding the
operation to the context, but then it would synchronously create the RCs
labels via network requests to consul. This means that if the
transaction fails later on, the labels will still be written resulting
in data inconsistencies. This commit changes CreateTxn() to add the
label creation operation to the transaction rather than applying them
immediately.

This also required changing the way that labels are applied to new RCs
even when NOT doing a transaction in order to maximize code reuse. The
non-transactional create function would loop over all label key sand set
them one by one (i.e. a separate network request per label key). This
won't work in a transaction because the full value of the label key
needs to be known when you set it and that value is cumulative. To use
transactions we have to compute the full set of labels first and then
apply it in a single operation in the transaction. It's better to do
this anyway to reduce the number of network operations to consul.

This further necessitated changing the Create() and CreateTxn() APIs so
that callers can supply additional labels. The roll store, for example,
allows the user to specify arbitrary labels that should be applied to
RCs that are created in addition to the default "pod_id" label. For the
same reasons stated above, when using a transaction these label changes
have to all be in the same operation. As a result, Create() and
CreateTxn() now have a new klabels.Set argument specifying additional
labels to apply when the RC is created, and those will be bundled into
the single KV operation that creates the labels.