Use a real consul store instead of a fake one in daemon set tests.
The hope is that this will make the tests less flakey. This is likely
because the fake daemon set store was attempting to implement watch-like
behavior using condition variables which is an error prone thing to do.

Now that we've crossed the bridge into using real consul instances in
our tests, we may as well do the same here.

Additionally this move makes sense because the tests were using a daemon
set store function to create daemon sets that doesn't exist anymore (now
there's only CreateTxn).