Delete unused mock daemon set store code.
This store used to be used with the goal of avoiding needing a consul
instance to run the tests. However it is suspected that the complex mock
code was the source of test flakes, and we've also made a broader push
toward using real consul stores as we adopt consul transactions.