roll farm: delete old RC in a transaction with deleting the RU
When a rolling update is complete, the old RC is typically deleted after
its replica count has been decremented to zero. Subsequently the RU is
deleted, signifying the end of the rolling update.

If for some reason the old RC is deleted and the RU is not, for instance
if the server is shut down in between or if the consul cluster becomes
unavailable, another farm instance will attempt to continue the rolling
update and be unable to do so since the old RC doesn't exist.

This commit bundles deleting the old RC and the RU itself into a single
transaction so this data inconsistency can't happen.