Fix test flake when using consulutil.WatchDiff()
There was a possible panic for sending on a closed channel by WatchDiff,
because the passed in error channel which gets written to whenever a
consul error is encountered might be closed by the calling code.

This commit makes it the responsibility of WatchDiff() itself to close
the channel when its work is done which safely prevents sending on the
channel after it is closed.