Expose GetLabelsWithIndex() on consul and http labels applicators.
This operation will be useful for building transactions which require an
index for check-and-set operations. Committing a transaction requires
direct access to consul, however we don't want building a transaction to
require that. However the HTTP applicator has been unable to build
certain transactions because it cannot get the modify index of a key via
the HTTP API.

This commit exposes modify index in the GetLabels handler in the labels
http server which allows the http applicator to expose it which will aid
in building transactions.