remove cancel func from transaction.Commit()
The cancel func was previously required because context cancelation was
used to guarantee that the same transaction is not applied twice.
However this makes the API cumbsersome and isn't really a proper use of
contexts.

This commit changes the underlying transaction data structure so that it
tracks a bool protected by a mutex that indicates whether the
transaction has already been applied. This removes the need for the
cancellation function.