Propagate transaction failure information from Commit()
When committing a transaction, there are 3 possible outcomes:
1) The transaction is successfully applied
2) The transaction is successfully "considered" by the server, but
rolled back due to a failing condition in any of the operations
contained in the transaction
3) The transaction is not successfully "considered" by the server, e.g.
due to a TCP error or temporary server unavailability.

There are cases where 2) and 3) might be handled differently, but prior
to this commit it was impossible to discern which scenario occurred from
Commit().

This commit changes Commit() to propagate a bool that indicates that the
transaction was not rolled back in addition to err. A new function
called MustCommit() was added as a convenience function for callers that
don't care why the transaction wasn't applied, e.g. in tests or CLIs
that are easy to retry.