Support "branching" consul transactions off of each other.
Prior to this commit, when creating a context containing a transaction
value with transaction.New(), an empty transaction would be stored in
the context regardless of the state of the transaction passed in.

This commit changes that behavior so that the context passed to New() is
inspected for a transaction, and if one exists then its operations will
be copied into the new transaction being created.

This is useful, for example, when you want to enforce that certain locks
are held before performing a sequence of actions. You can make a single
transaction that has the required "check-session" operations, and then
create a bunch of transactions that "branch" of of that one and save the
work of rebuilding the "check-session" operations every time.

In particular this will be useful in the rolling update Run() function
which should only perform consul operations while it holds RC locks. If
the locks are lost, we want any operations to be rolled back and for the
loop to exit.