Add LockTxn function to consulutil.Session.
This function takes 3 context as arguments as well as a key to lock with
the session on which LockTxn() is being called.

The first context called the "lockCtx" will have the required kv
operations added to lock the key specified. The caller must commit the
transaction before any locks are acquired.

The second context is called the "unlockCtx" which will have the
required operations to unlock the key specified. The caller must commit
the transaction before the lock is released, and it will fail if the
lock isn't actually held by the session when the transaction is
committed.

The third context is called the "checkLockCtx" which will have the
operations required to test that the session holds the lock on key. This
is useful for callers that want to make consul operations that will only
succeed if the lock is still held. For example, the RU farm might want
to only mutate an RC if it holds the mutation lock for that RC.