Add CommitWithRetries() convenience function to transaction package.
CommitWithRetries() will retry failed commits indefinitely until either
the commit succeeds or the passed context is canceled. It will exit as
soon as a transaction is successfully applied whether or not that
transaction was rolled back (because in most cases it doesn't make sense
to retry a transaction that was rolled back).

It uses an exponential backoff strategy to avoid bogging down the consul
server.