Add EnableTxn() and DisableTxn() to RC store.
Additionally this commit consolidates the code used by the DeleteTxn()
function with EnableTxn() and DisableTxn(). It also shares code between
retryMutate() and mutateRCTxn()

These transactions will be useful for the roll farm to enable the new RC
and disable the old one in a transaction and additionally to ensure that
locks are held while this action is performed.