Make rolling updates use contexts and consul transactions.
Rolling updates previously worked as follows:

1) Farms compete with each other to lock new RU records. The winner then
proceeds to handle the update while the losers give up.
2) The update loop acquires mutation locks on the new and old RCs in
order to prevent race conditions where multiple update loops process the
same RU and also prevents other tools from changing RC counts from
underneath the RU.
3) The update loop will process the update until either a) it is
completed or b) it gets a cancellation signal, which only occurs if the
consul session used to grab the locks dies (and thus the locks are
released).

This is a pretty good approach, but if operations are undertaken by the
roll loop in the period of time after the session dies but before the
cancellation signal is noticed, there can be an overlap where multiple
RU farms are processing the same RU.

This commit addresses that issue by using consul transactions for every
consul operation the RU loop performs and also ensuring that those
transacitons include "check-session" operations on the RC locks. This
means that even if the cancellation signal has not been noticed but the
locks are gone, the transaction will fail and the RU loop can exit.

Additionally using transactions prevents other data inconsistency
conditions, for instance where the old RC is deleted but not the RU
record. This issue has been problematic since it requires special cases
in roll loop initialization to correct. Eliminating the possibility of
this condition is better.