Make consul aggregator aggregation rate configurable.
Consul aggregators are useful when selectors are going to be resolved at
a high rate yet it is desirable to limit the rate of requests to the
consul data store by trading latency. Prior to this commit the maximum
request rate would be one every 10 seconds, but in tests we might
require a shorter period of time to allow tests to complete in a
reasonable time frame.