Improve fake preparer so it makes fewer consul operations
The fake preparer is useful for testing code that interacts with the
preparer, for example daemon sets. The fake preparer just copies
manifests from intent into reality and deletes manifests from reality
when they're deleted from intent.

This commit makes an optimization to the fake preparer where it will
only write a manifest to reality if it differs from the one in intent
(before it did this unconditionally). This reduces the amount of log
output that appears in tests due to consul operations (and might improve
test times by putting less load on consul).