Fix an inverted for loop exit condition.
After the preparer launches a oneoff pod, it writes a status record for
that pod indicating that it has launched. It is intended to retry
forever until the record is written successfully.

However a typo in the for loop condition did the opposite, it will retry
forever if it succeeds and only exit when an error is hit (even if there
was no successful write).