Acquire a lock to synchronize creation of config for a pod cluster.
When creating or updating config for a pod cluster, the process was:

1) do a label lookup to see if a config already exists for the pod
cluster
2a) if it doesn't, generate an ID for the config and create it
2b) if it does, use the existing ID and update it

This had a race condition where two separate processes doing this
operation might both perform step 2a.

This commit starts to mitigate the problem by first acquiring a lock
namespaced by the pod cluster before performing the label lookup which
is only released after the record is written.

This does not fully address the problem because it's possible that the
lock could be lost before the write is performed. A followup commit will
perform the write within a transaction that tests that the lock key is
still held by the session.