rc: send an alert if any "current" pods are on "ineligible" nodes
The main RC processing loop is designed in a paranoid fashion in which
it will only schedule nodes if the replica count becomes larger than its
current membership and it will only unschedule nodes if the replica
count becomes smaller than its current membership.

Node eligibility was only considered when unscheduling or scheduling.

This means that situations can arise where the nodes on which the RCs
pods are scheduled might become "ineligible" and the RC won't take any
action, and human operators might not notice for a long period of time.

This commit makes the RC consider the set of eligible nodes every loop
and send an alert if any of the current pods are scheduled on ineligible
nodes. It does not attempt to take any corrective action because this
likely requires input from human operators about how to safely to so.