Treat hook failures like normal failures.
Up until this commit, hook failures woul not stop a pod from being
installed and having reality updated. Hook failures were logged but
otherwise ignored.

This commit changes this by treating a hook failure as a critical
failure which will cause the entire pod installation routine to be
retried (the same way a download failure is treated, for example).

This means that P2 operators will have confidence that if reality is
updated for a pod, that all of its hooks must have succeeded. Previously
this was not the case and the philosophy was that P2 applications should
check for conditions they expect hooks to put in place within their
health checks, which will prevent a rolling update from rendering all
pod instances within a pod cluster to be nonfunctional