extend replicator to allow transactional application of pod labels
Prior to this commit, daemon sets would label the pods they schedule
ahead of time (with the daemon_set_id label). However, it's preferable
to schedule a pod and label it in the same consul transaction, so that
you can tell just by looking at a pod's labels whether it has been
scheduled by the daemon set or not.

In order for this to occur, the replication package had to be extended
to allow providing a set of "pod labels" that should be applied to each
pod scheduled by the replication.

This commit adds that functionality and makes the daemon set farm use it
instead of scheduling pods itself.

This commit attempts to make the bare minimum architectural changes to
daemon sets.