Don't try to add user in hook in travis integration tests
The travis environemnt does not have the /sbin/adduser command
available, instead we add the users the test needs in travis.sh.

This commit makes sure we invoke the integration tests with a benign
hook that won't fail. This matters because the preparer no longer
ignores hook failures and treats them as critical failures