Use contexts for cancellation in daemon set farm.
This will be useful in order to pass a context containing a partial
consul transaction that ensures that the daemon set farm still holds the
lock that it's supposed to be when it performs consul operations. This
will be useful to protect against multiple daemon set farms writing the
same daemon set status entry with conflicting information.