Improve transactional locking API
This commit reworks the way LockTxn() works on a consul session. Prior
to this commit you had to pass in 3 different context.Context values
each with an inner consul transaction, and LockTxn() would load each of
them up with the respective consul operations that enable locking,
unlocking, and checking that the lock is held within consul
transactions.

In practice this is very cumbersome since callers acquiring locks must
prepare all of the contexts ahead of time, and preserve the contexts and
their cancellations everywhere where they are needed.

This commit changes the LockTxn() function to only take a single context
intended for locking the key and returns an interface which can be used
to load up contexts with unlocking and check-locking operations
respectively.

This will means callers can simply pass around a reference to that
interface and then create contexts on the fly when needed for unlocking
or checking that a lock is held within a transaction.