Fix a daemon set farm test that was flakey
There are some places in the daemon set tests where it will make a
mutation to a daemon set and then repeatedly poll the daemon set farm's
internal data structures to make sure that the daemon set mutation is
reflected.

This commit found a place where the condition was set up improperly to
look for the the daemon set BEFORE the mutation was applied, which
sometimes would succeed if the poll occurred fast enough that the
internal data structure didn't yet reflect the update. This commit fixes
that condition to work properly