pods: Pass fetcher from factory to pods
NewFactory was taking a fetcher and then not using it, instead using a
defaultFetcher.

This is inappropriate for situations where we must configure specific
settings on the fetcher in order to be able to download artifacts (for
example, CA file)

`newPodWithHome` must accept a fetcher and then use it, and `NewUUIDPod`
and `NewLegacyPod` must pass the fetcher given from `NewFactory`.

Recall that the certificates we issue require that we be able to
configure specific CAs.

This was partially done in #826, but it was not made to apply to Pods,
which internally contain Fetchers as well.