fix potential panic in pkg/ds/farm_test.go
The panic happened because t.Error() might be called from a goroutine
spawned by newTestSession() after the test function has exited. This is
fixed by adding a defer session.Destroy() call in the test function
which guarantees that the renewal error channel will be closed before
the function exits, which makes it impossible for an error to come down
that channel.