Fix cancellation of CommitWithRetries()
The intention of CommitWithRetries() is that it will do retries with
exponential backoff of the transaction until it either succeeds or the
context is canceled. However the context cancellation wasn't working
properly because it used a "break" inside of a select statement when the
intention was to break out of the for loop.

A test was written that was intended to catch the bug, but the test
didn't because the passed transaction had no operations on it, which
means that Txn() wouldn't even be called on the txner, therefore
bypassing the configured error.