Add ability to check if a replication is in progress
When the daemon set farm is reporting status for a daemon set, it was
determining if a replication in progress by checking if the
currentReplication variable is set. This didn't quite work because the
currentReplication variable isn't unset when a replication completes,
and we don't want to unset it because the status routine should still be
able to ask for the completed count.

This commit adds an InProgress() call to the replication interface which
will answer whether the replication is in progress or not which will be
used to determine the replication_in_progress field in the daemon set
status struct