Add random jitter to consul watch requests after an error.
Many places in p2 initiate "watches" on the consul database, meaning
they will have many long-lived HTTP connections that are re-initiated as
soon as the response is received.

This causes a lot of synchronized load on consul whenever a node becomes
unavailable (whether due to hardware failure or a graceful restart). The
watches that that node is handling will all be broken at the same time
and then re-attempted all at the same time, and often the retries will
sync up as well since the same wait time on consul is used (usually 5
minutes).

This commit adds the ability to specify a jitter window from which a
random value between 0 and the jitter window will be chosen and slept
before intiating another request after an error. This should smooth out
the load on consul after a server is rebooted.