refactor daemon sets to obviate need for halting replications
Previously daemon sets would operate by scheduling replications that
might contain thousands of nodes, which can take on the order of hours.
This posed a problem because daemon sets occasionally need to add new
nodes or remove nodes based on node selector changes. This required
halting the replication and starting another one.

Under certain circumstances this means a replication is never allowed
to complete because there are many circumstances that can cause the
cancellation.

This commit makes the following changes:

1) It's possible to create a replication that reads nodes to deploy off
of a channel rather than requiring a list up front. This allows daemon
sets to start a single replication and reuse it over the lifetime of the
daemon set.
2) Removing pods from a daemon set does not stop replication. Currently
there's no functionality to guarantee that a node being removed isn't
scheduled again, but that will "eventually" happen because if the
scheduling routine writes it again, the daemon set will just remove it
again. In other words, the same node might need to be removed multiple
times before the change is permanent.
3) Adding pods just puts the new nodes on the channel being read by the
replication. This should hopefully result in fast deploys to new nodes
if the replication channel is being processed fast enough.
4) Whenever the daemon set's manifest changes, an atomic operation is
made on the replication that causes it to start using the new manifest
right away. The daemon set will then add the full node list to the
channel to ensure that every node gets the manifest change.

Performance concerns:

There is a buffered channel that the so-called "WatchDesires" loop
passes newly added nodes on to the "Replication" goroutine. The
replication goroutine pops nodes off of the buffered channel and feeds
each one to the internal goroutine to the replicaiton package. This can
take hours which is why a buffered goroutine is used to allow the
"WatchDesires loop" to still respond quickly to other changes like
manifest changes or disables.

It's possible that this channel could fill up and cause daemon sets to
become less responsive to updates as they block sending on the channel.

For now this risk is deemed acceptable because deploys are expected to
be infrequent, and it shouldn't be a problem at all to keep up with node
changes.