Use a new session for each RU to acquire RC locks.
This simplifies cleanup at the end of the rolling update handler, mainly
that we want to unconditionally release RC locks even if an error or
other cancellation is encountered. This is a good use case for a new
session because the handler should release locks without needing the
whole farm to shut down regardless of the kind of error that occurs.

This also avoids the need for handling the case where the root context
is canceled because the RU gets deleted which might cause transactional
operations to fail since they too depend on the root context. Even if
the transactions are canceled we can still destroy the session to
release the locks.