pkg/alerting: support different pagerduty urgency levels in alerts
The alerting package provides an interface for sending alerts when
something goes wrong within P2, and has an implementation of that
interface that uses the pagerduty API.

This commit adds the ability to configure the pagerduty alerter with
high and low urgency service keys to use with pagerduty's API as well as
the ability to specify whether an alert is high or low urgency. The
correct key will be used depending on the urgency of the alert.