Add a http timeout of 6 minutes to consul clients
We expect that responses might take up to 5 minutes (the wait time we
specify for consul operations) to return, but any longer than that is an
indication that the request isn't going to return. This commit sets a
timeout of 6 minutes on http requests to consul so that they will be
retried.