Add LockIfKeyNotExists() function
This will be used by the replication controller to schedule a new node
during a node transfer. It guarantees that the key did not previously
exist. If it fails, a new RC has scheduled the node and the old RC knows
to abort the node transfer. Also, if the node transfer routine dies, the
session will die and the key will be cleaned up automatically.