Fix an issue with daemon set responsiveness
Commit 643738a9006582eb9d4df9f9469f26be5b97a188 attempted to collapse a
lot of duplicated code into a single function in the daemon set farm,
however an unintended side effect was that the daemon set farm is slow
to pass updates to a daemon set to the goroutine that is responsible for
carrying out the update.

The reason is that daemon set contention checks were being performed
regardless of whether the daemon set had been updated, so every time the
daemon set watch would fire a contention check would be run against
every daemon set. This is an expensive calculation because it requires
performing label select operations for every daemon set and then looking
for overlaps between sets.

This commit performs the check only if the daemon set is new or has been
updated.