Add rest of rc node transfer functionality
This change adds the remaining pieces to allow replication controllers
with the cattle allocation strategy to automatically schedule a new node
and unschedule an old node when a node is discovered to be ineligible.
It adds a way to schedule with a session, a goroutine to watch for node
health, and a function to label the new node and unschedule to old. It
is also designed so that the transfer can be rolled back if a deploy
happens before it is finished or if it errors in any way.

It rolls back by tying the scheduling of the new node to a session. The
key written to intent/ will be lost if the session ends so no record of
the transfer exists.