Update ensureConsistency to remove race
It used to be possible (albeit unlikely) for a node transfer to finish
before ensureConsistency was called. Therefore the old node would be
unscheduled but still be in the `current` slice. It would therefore be
scheduled again despite having just been intentionally removed.

This change checks for node eligibility in ensureConsistency to remove
that chance.