Reschedule new node if node transfer restarts
Before, if a node transfer were to restart, the new node would not be
rescheduled if a node transfer status existed in the store. A new node
would therefore not be scheduled because the session the node was
originally scheduled with would expire. This change calls
scheduleWithSession() every time the transferNodes function executes to
be sure that the new node is scheduled.