pkg/rc: stop using an RC pointer passed to a watch function
Prior to this commit, the main replication controller loop would start
by calling the Watch() function on the RC store, passing it a pointer to
an RC struct and also a mutex. Whenever the watch goroutine detected a
change to the RC, it would lock the mutex and then mutate the RC
pointer.

This meant that whenever the main RC loop tried to access a field on the
RC pointer, it would need to grab a mutex while it did so. This is
untenable as new references to the field are added all the time and it's
hard to remember to hold the mutex, creating data races.

Furthermore, it's hard to reason about what happens if the RC is mutated
in the middle of a meetDesires() invocation.

To address this, this PR changes the way Watch() on the rc store works
so that it returns a channel of RC structs (not pointers). Whenever a
value is read from the channel, this value is passed to meetDesires()
where it is processed without fear of mutation occurring during the
call.

Now the only mutex needed by the replication controller handler is to
guard the node transfer struct which is still used by multiple
goroutines.