ensure that the preparer is always using the latest pod manifest
The p2-preparer has a main goroutine that watches a key prefix in consul
and converts the data into pod manifests each time the watch fires
(either because a manifest changed or because the watch timeout was
hit). If no goroutine exists for a particular pod ID, the preparer will
spawn a goroutine dedicated to handling updates to the manifest for that
pod.

Each pod goroutine consumes manifests on a channel, and if the manifest
differs from the manifest in the reality tree then an update will occur.

In order to prevent blocking the main preparer goroutine on pods that
are slow to install (due to slow enable or disable scripts, or large
downloads for example), the channel is written within a select statement
that contains a default case. The thought was that the next time the
watch fires the manifest will be written again, which is only delay of a
few minutes at most.

However, if the pod goroutine is repeatedly busy doing a lot of work
such as installing a manifest with a missing launchable, the sending and
the receiving goroutine might miss each other, resulting in several
manifest updates occurring without the pod goroutine handling it.

This commit addresses the issue by buffering each pod's manifest
channel with a size of 1. This allows the sending goroutine to maintain
its not-blocking behavior but guarantees that the next read by the
receiving goroutine will get the latest manifest. The sender attempts to
drain the channel before writing a value, and logs a warning if the
manifest that was waiting on the channel is different than the one being
sent to indicate that the pod goroutine is consuming slowly.