Define new data type for node transfer audit logs
These audit log records will be written to consul by the RC farm as it
takes one of the three actions pertaining to node transfers: start one,
finish one, roll one back.

This will provide an opportunity for an external service to ingest these
audit log records and persist them to a database.