Call allocate before deallocate in node transfer
This change prevents the chance of a logical error in which
AllocateNodes() would return the same node that DeallocateNodes() just
removed. While that wouldn't happen if we set the cologroup of the deallocated
node correctly, it is nice to ensure it can't happen even in the event of
cologroup oversight.